requirements.txt
fastapi==0.135.1
filelock==3.25.1
langchain_mongodb==0.11.0
langchain_voyageai==0.3.2
loguru==0.7.3
protobuf==7.34.0
pydantic==2.12.5
pydantic_settings==2.13.1
pymongo==4.16.0
python-dotenv==1.2.2
python_jose==3.4.0
starlette==0.52.1
tenacity==9.1.2
uvicorn==0.41.0
websockets==13.1
google-genai

